Support status clearing by passing on empty emoji string
Slack announced expiring statuses today, which is pretty cool. I've
definitely left "out for lunch" as my status for hours after getting
back.

But I realized that a definite deadline doesn't really match my
habits. So instead I went to check how to clear a status, and it turns
out this is all that was required.

In retrospect, ":" does satisfy both "starts with" and "ends with",
but wasn't quite what I had in mind :)